# Conflicts:
#   pubspec.lock

+++ lib/presentation/widgets/approval_pending_advocate_card.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:jusitfi_admin/presentation/widgets/rating_box.dart';
-import 'package:jusitfi_admin/presentation/widgets/starrating.dart';
 import 'package:jusitfi_admin/presentation/widgets/videoCallWidget.dart';
 import 'package:jusitfi_admin/presentation/widgets/view_cancel_reason_dialog.dart';
 import 'package:jusitfi_admin/utils/constants/colors.dart';
-import 'package:smooth_star_rating_null_safety/smooth_star_rating_null_safety.dart';
 
 import '../../utils/constants/textstyles.dart';
 import 'advocate_profile_table.dart';
@@ -76,7 +74,7 @@ class ApprovalPendingAdvocateCard extends StatelessWidget {
                     )
                   ],
                 ),
-                RatingBox(rating: rating),
+                RatingBox(rating: rating.toInt()),
               ],
             ),
             const SizedBox(height: 8),
